Initial idea for terminal assistant
The terminal assistant is a LLM assistant that helps with Terminal
commands. Mainly useful for complex and once-in-a-lifetime commands, so
we don't need to switch to browser to search the exact command to do
something.

It uses a API key from Deepseek to generate the messages. I tried some
local LLM models, but none could properly obey the instructions to make
the answer short and precise, so I decided for Deepseek as it was able
to comply with that request and being relatevely cheap.

Future ideas are to switch to C++, and execute the given command with a
simple ENTER key press.

Signed-off-by: Lucas Cardoso <lucascardoso910@gmail.com>